chore(Dockerfile): Move abstruse necessary Dockerfile commands
These commands don't need to be at the end of the file as Docker doesn't actually care.

Doing this opens up the possibility of using a custom build image and 'FROM'ing from that. Perhaps abstruse could build that image as a part of the install and the "images" section of the site could literally be an area to add customisations. i.e.

    FROM abstruse-build-image
    # add custom things

    # example; install sqlite3
    RUN sudo apt-get install sqlite3 -y